website_dashboard/requirements.txt
flask==2.3.3
pandas==2.0.3
fastapi==0.104.1
uvicorn==0.23.2
wandb==0.15.12
requests==2.31.0
python-dotenv==1.0.0